Fix ownership-check type coercion and trim vendor bloat in release build
Cast candidate IDs to int before comparing in the frater ownership
guards, since SQLite/PDO don't reliably return the same PHP type on
both sides of a strict comparison. Also strip non-runtime files
(tests, docs, tools, .git) from vendor/ after composer install, since
this environment's proxy blocks dist-zip downloads and falls back to
full git-source checkouts — cut the release zip from ~270MB to ~18MB.

+++ app/Http/Controllers/Frater/ScheduleItemController.php
@@ -22,7 +22,7 @@ class ScheduleItemController extends Controller
 
     public function update(Request $request, ScheduleItem $scheduleItem): RedirectResponse
     {
-        abort_unless($scheduleItem->candidate_id === Auth::guard('frater')->id(), 403);
+        abort_unless((int) $scheduleItem->candidate_id === (int) Auth::guard('frater')->id(), 403);
 
         $scheduleItem->update($this->validated($request));
 
@@ -32,7 +32,7 @@ class ScheduleItemController extends Controller
 
     public function destroy(ScheduleItem $scheduleItem): RedirectResponse
     {
-        abort_unless($scheduleItem->candidate_id === Auth::guard('frater')->id(), 403);
+        abort_unless((int) $scheduleItem->candidate_id === (int) Auth::guard('frater')->id(), 403);
 
         $scheduleItem->delete();
